+++ NEWS.md
@@ -1,5 +1,7 @@
 # angstroms dev
 
+* added `databoundary` function, to trace around valid pixels
+
 * more general approach to reading slices, with `romsdata3d` and `romsdata2d`
 
 * S3 methods for rawdata